+++ regex/src/com.oracle.truffle.regex.test/src/com/oracle/truffle/regex/tregex/test/PythonByteTests.java
@@ -46,7 +46,7 @@ public class PythonByteTests extends RegexTestBase {
 
     @Override
     String getEngineOptions() {
-        return "Flavor=PythonBytes";
+        return "Flavor=Python,Encoding=LATIN-1";
     }
 
     @Test